Persist last kid record
The baby connect service is intended for synchronizing data. When you request
the status list, you are supposed to specify the last record received for
that kid. In the result "list" it will include new entries. You should update
the last record for this kid so that you can specify it in the next request.